fix(benchmark): compile capacity suite on Windows
Keep the operational file lock on Linux and provide a compile-only unsupported implementation elsewhere. The capacity suite remains runtime-limited to linux/amd64.
